+++ CHANGELOG.md
@@ -1,10 +1,12 @@
-
 # Change Log
 
 ## 0.1.3
 
 - Added 'org.eclipse.swt.dnd' to glimmer auto-included Java packages
 - Updated Tic Tac Toe sample to use new `message_box` keyword 
+- Add DragSource and DropTarget transfer expression that takes a symbol or symbol array representing one or more of the following: FileTransfer, HTMLTransfer, ImageTransfer, RTFTransfer, TextTransfer, URLTransfer
+- Set default style DND::DROP_COPY in DragSource and DropTarget widgets
+- Support Glimmer::SWT::DNDProxy for handling Drop & Drop styles
 
 ## 0.1.2